refactor(renderer): split Shell/ChangesPanel, bundle thread-row actions, canonical helpers
quality-review slice 5 (5a, 5b, 5e, 5f).

5a: split the Projects subtree (WorkspaceNav / ProjectRow / ArchivedSection /
NavThread / threadLabel + their constants) out of Shell.tsx (987→291) into
shell/workspace-nav/. The four per-row callbacks (select / delete / setThreadFlags /
rename) plus new-thread-in-workspace / remove-workspace are bundled into ONE explicit
`ThreadRowActions` prop threaded Shell → WorkspaceNav → ProjectRow →
NavThread/ArchivedSection (no context). App passes the bundled `actions` object.

5b: extract git/BranchMenu.tsx, git/PrSection.tsx (with PrChip + prStateClass), and
git/FileRow.tsx out of ChangesPanel.tsx (806→335, keeps the status subscription,
commit state, and layout). `glyphClass` moves to git/status-view.ts next to fileGlyph
(with a test case) and FileRow consumes it from there.

5e: extract `menuItemClass` + `menuSurfaceClass` into ui/menu-styles.ts, composed via
cn() in menu / context-menu / popover / select. The item classes stay byte-identical
per component (context-menu keeps its data-[disabled] locally); the surface tokens are
shared (padding/sizing/z stay local — same rendered utilities).

5f: one `basename()` in lib/paths.ts replaces basenameOf (file-link),
fileBasename (breadcrumb-segments), and the inline slice in FilePreview; and the 11
side-panel-store `xWorkspaceY` wrappers collapse behind a typed `bindWorkspaceOp`
factory (call sites and store semantics unchanged).
