+++ src/renderer/src/git/ChangesPanel.tsx
@@ -1,20 +1,12 @@
 import { useEffect, useState, type JSX } from 'react'
-import {
-  Boxes,
-  Check,
-  ChevronDown,
-  GitBranch,
-  GitCommitHorizontal,
-  GitPullRequest,
-  Monitor,
-  PanelRightClose,
-  RefreshCw,
-} from 'lucide-react'
-import type { GhPr, GhPrResult, GitBranch as GitBranchInfo, GitStatus } from '../../../shared/ipc'
-import { cn } from '../lib/utils'
-import { Badge, Button, IconButton, Input, Menu, MenuContent, MenuItem, MenuSeparator, MenuTrigger, Textarea } from '../ui'
+import { Boxes, GitCommitHorizontal, Monitor, PanelRightClose, RefreshCw } from 'lucide-react'
+import type { GitStatus } from '../../../shared/ipc'
+import { Badge, Button, IconButton, Textarea } from '../ui'
 import { getCommitDraft, setCommitDraft } from './commit-draft-store'
-import { buildChangesView, reconcileUnchecked, type GitFileView } from './status-view'
+import { buildChangesView, reconcileUnchecked } from './status-view'
+import { BranchMenu } from './BranchMenu'
+import { PrSection } from './PrSection'
+import { FileRow } from './FileRow'
 import { DiffWorkerProvider } from './DiffWorkerProvider'
 import { DiffView } from './DiffView'
 
@@ -339,468 +331,5 @@ function EnvPlaceholder({ icon, children }: { icon: JSX.Element; children: strin
       <span className="min-w-0 flex-1 truncate">{children}</span>
       <span className="shrink-0 text-[10px] font-medium text-faint">Soon</span>
     </div>
-  )
-}
-
-/**
- * The branch header dropdown (#87): the current branch name as a base-ui Menu trigger
- * that, ON OPEN, fetches the Workspace's branches (`gitBranches`) and lists them for a
- * one-click checkout (`gitCheckout`); a "Create branch…" item reveals an inline input
- * (`gitCreateBranch`). The ahead/behind indicator stays beside the trigger (unchanged).
- *
- * Disabled while `busy` (a turn streams) — the same guard as commit: we don't switch
- * branches under the agent mid-turn (a switch rewrites the working tree). A checkout /
- * create error surfaces inline + recoverable (the common one is git's dirty-tree
- * refusal — NO data loss, git protects). The status STREAM updates the header on a
- * successful switch (main re-reads after the op), so this holds no branch-name state.
- */
-function BranchMenu({
-  workspaceDir,
-  branch,
-  ahead,
-  behind,
-  busy,
-}: {
-  workspaceDir: string
-  branch: string
-  ahead: number
-  behind: number
-  busy: boolean
-}): JSX.Element {
-  const [branches, setBranches] = useState<GitBranchInfo[] | null>(null)
-  const [listError, setListError] = useState<string | null>(null)
-  const [loading, setLoading] = useState(false)
-  // The recoverable error from a checkout / create (dirty-tree refusal, name collision).
-  const [opError, setOpError] = useState<string | null>(null)
-  // The inline "Create branch…" affordance: revealed by the menu item, below the header.
-  const [creating, setCreating] = useState(false)
-  const [newName, setNewName] = useState('')
-  const [busyOp, setBusyOp] = useState(false)
-
-  // Fetch the branch list each time the menu opens, so it reflects branches created out
-  // of band (a terminal `git branch`, the agent) without a panel-wide subscription.
-  async function loadBranches(): Promise<void> {
-    setLoading(true)
-    setListError(null)
-    try {
-      const result = await window.api.gitBranches({ workspaceDir })
-      if (result.ok) setBranches(result.branches)
-      else {
-        setBranches([])
-        setListError(result.error)
-      }
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  async function checkout(name: string): Promise<void> {
-    if (busy || busyOp) return
-    setBusyOp(true)
-    setOpError(null)
-    try {
-      // Pass the branch's FULL name + `track`: a remote-only `<remote>/<branch>` goes
-      // through `git switch --track` (an unambiguous tracking-local create — robust even
-      // with two remotes sharing a trailing name); a local name (which may contain `/`,
-      // e.g. `feat/x`) switches verbatim. `info.isRemote` decides.
-      const info = branches?.find((b) => b.name === name)
-      const result = await window.api.gitCheckout({ workspaceDir, name, track: info?.isRemote ?? false })
-      // On success the streamed status refresh updates the header to the new branch.
-      if (!result.ok) setOpError(result.error)
-    } finally {
-      setBusyOp(false)
-    }
-  }
-
-  async function createBranch(): Promise<void> {
-    const name = newName.trim()
-    if (!name || busy || busyOp) return
-    setBusyOp(true)
-    setOpError(null)
-    try {
-      const result = await window.api.gitCreateBranch({ workspaceDir, name })
-      if (result.ok) {
-        // The header updates via the streamed status refresh; clear + close the input.
-        setNewName('')
-        setCreating(false)
-      } else {
-        setOpError(result.error)
-      }
-    } finally {
-      setBusyOp(false)
-    }
-  }
-
-  return (
-    <>
-      <div className="flex items-center gap-1.5 border-b border-border-muted px-3 py-2 text-[13px] text-muted">
-        <Menu
-          onOpenChange={(open) => {
-            if (open) void loadBranches()
-          }}
-        >
-          <MenuTrigger
-            disabled={busy}
-            title={busy ? 'Agent is working…' : branch}
-            className={cn(
-              'flex min-w-0 flex-1 items-center gap-1.5 rounded-md px-2 py-1 text-left transition-colors',
-              'hover:bg-accent/10 disabled:cursor-not-allowed disabled:opacity-50 disabled:hover:bg-transparent',
-            )}
-          >
-            <GitBranch size={14} aria-hidden className="shrink-0" />
-            <span className="min-w-0 flex-1 truncate font-medium text-text">{branch}</span>
-            <ChevronDown size={13} aria-hidden className="shrink-0" />
-          </MenuTrigger>
-          <MenuContent align="start" className="max-h-80 min-w-48 overflow-y-auto">
-            {loading ? (
-              <div className="px-3 py-1.5 text-sm text-muted">Loading…</div>
-            ) : listError ? (
-              <div className="px-3 py-1.5 text-sm text-bad">{listError}</div>
-            ) : branches && branches.length > 0 ? (
-              branches.map((b) => (
-                <MenuItem key={b.name} onClick={() => void checkout(b.name)} disabled={b.current}>
-                  <Check size={13} aria-hidden className={cn('shrink-0', b.current ? 'opacity-100' : 'opacity-0')} />
-                  <span className="min-w-0 flex-1 truncate">{b.name}</span>
-                  {b.isRemote && <span className="shrink-0 text-[10px] uppercase text-muted">remote</span>}
-                </MenuItem>
-              ))
-            ) : (
-              <div className="px-3 py-1.5 text-sm text-muted">No branches.</div>
-            )}
-            <MenuSeparator className="my-1 h-px bg-border-muted" />
-            <MenuItem onClick={() => setCreating(true)}>
-              <span className="w-3 shrink-0" aria-hidden />
-              Create branch…
-            </MenuItem>
-          </MenuContent>
-        </Menu>
-        {(ahead > 0 || behind > 0) && (
-          <span className="shrink-0 tabular-nums" title={`${ahead} ahead, ${behind} behind`}>
-            {ahead > 0 && <>↑{ahead}</>}
-            {ahead > 0 && behind > 0 && ' '}
-            {behind > 0 && <>↓{behind}</>}
-          </span>
-        )}
-      </div>
-
-      {creating && (
-        <div className="flex items-center gap-1.5 border-b border-border-muted px-3 py-2">
-          <Input
-            autoFocus
-            aria-label="New branch name"
-            value={newName}
-            onChange={(e) => setNewName(e.target.value)}
-            placeholder="New branch name"
-            disabled={busy || busyOp}
-            className="h-8 min-w-0 flex-1 text-[13px]"
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') {
-                e.preventDefault()
-                void createBranch()
-              } else if (e.key === 'Escape') {
-                e.preventDefault()
-                setCreating(false)
-                setNewName('')
-                setOpError(null)
-              }
-            }}
-          />
-          <Button
-            type="button"
-            size="sm"
-            onClick={() => void createBranch()}
-            disabled={busy || busyOp || newName.trim().length === 0}
-          >
-            Create
-          </Button>
-        </div>
-      )}
-
-      {opError && (
-        <p className="border-b border-border-muted px-3 py-2 text-[11px] text-bad" role="alert">
-          {opError}
-        </p>
-      )}
-    </>
-  )
-}
-
-/**
- * The PR chip / Create-PR affordance (#88, slice 4) below the branch header. On a branch
- * change (and a manual refresh, via `refreshKey`) it fetches the current branch's GitHub
- * PR through `gh` (`ghCurrentPr` — a NETWORK call, so NOT tied to every status tick):
- *  - a PR exists → a compact, state-tinted chip rendered as an EXTERNAL anchor
- *    (`target="_blank"`): clicking it routes through main's `setWindowOpenHandler` ->
- *    `shell.openExternal`, opening the PR in the system browser (no new IPC).
- *  - no PR (and the branch is sensible — not detached, not the repo's default) → a
- *    "Create PR" affordance: a small title+body form -> `ghCreatePr`. GATED on the branch
- *    having an upstream (`hasUpstream`): `gh pr create` non-interactively can't prompt to
- *    push, so without an upstream we show "Push your branch first" instead of failing. The
- *    form is DISABLED while `busy` (a turn streams) — the same guard as commit/branch ops.
- *  - gh missing / not authed (`{ok:false}`) → a subtle muted hint, never a crash.
- * The default-branch check is a best-effort `gitBranches` probe (local, no network); when
- * it can't resolve a default, we simply allow Create-PR (gh would surface any real error).
- */
-function PrSection({
-  workspaceDir,
-  branch,
-  detached,
-  hasUpstream,
-  busy,
-  refreshKey,
-}: {
-  workspaceDir: string
-  branch: string
-  detached: boolean
-  hasUpstream: boolean
-  busy: boolean
-  refreshKey: number
-}): JSX.Element | null {
-  // `'loading'` before the first fetch resolves; then the `GhPrResult`. Null only while
-  // detached (no PR surface). The created-PR chip lands via a re-fetch after a create.
-  const [result, setResult] = useState<GhPrResult | 'loading' | null>('loading')
-  const [isDefault, setIsDefault] = useState(false)
-  const [creating, setCreating] = useState(false)
-  const [title, setTitle] = useState('')
-  const [body, setBody] = useState('')
-  const [submitting, setSubmitting] = useState(false)
-  const [createError, setCreateError] = useState<string | null>(null)
-
-  useEffect(() => {
-    if (detached) {
-      setResult(null)
-      return
-    }
-    let cancelled = false
-    setResult('loading')
-    setCreating(false)
-    setCreateError(null)
-    void window.api.ghCurrentPr({ workspaceDir }).then((res) => {
-      if (!cancelled) setResult(res)
-    })
-    // Best-effort default-branch probe (local git, no network): suppress Create-PR on the
-    // repo's default branch. A failed/unresolved probe leaves `isDefault:false` (allow).
-    void window.api.gitBranches({ workspaceDir }).then((res) => {
-      if (cancelled) return
-      setIsDefault(res.ok ? (res.branches.find((b) => b.current)?.isDefault ?? false) : false)
-    })
-    return () => {
-      cancelled = true
-    }
-  }, [workspaceDir, branch, detached, refreshKey])
-
-  async function createPr(): Promise<void> {
-    const t = title.trim()
-    if (!t || busy || submitting) return
-    setSubmitting(true)
-    setCreateError(null)
-    try {
-      const res = await window.api.ghCreatePr({ workspaceDir, title: t, body: body.trim() })
-      if (res.ok) {
-        // Re-fetch so the freshly-created PR renders as the real chip (number/title/state);
-        // the form closes. gh already printed the URL — the chip's anchor opens it.
-        setCreating(false)
-        setResult('loading')
-        void window.api.ghCurrentPr({ workspaceDir }).then(setResult)
-      } else {
-        setCreateError(res.error)
-      }
-    } finally {
-      setSubmitting(false)
-    }
-  }
-
-  // Detached HEAD / before the first fetch: render nothing (no PR surface).
-  if (detached || result === null) return null
-  if (result === 'loading') {
-    return <p className="border-b border-border-muted px-3 py-1.5 text-[11px] text-muted">Checking pull request…</p>
-  }
-  // gh missing / not authed / an unexpected gh failure: a subtle hint, not a crash.
-  if (!result.ok) {
-    return (
-      <p className="border-b border-border-muted px-3 py-1.5 text-[11px] text-muted" title={result.error}>
-        {result.error}
-      </p>
-    )
-  }
-  // A PR exists → the state-tinted chip, opened externally via target="_blank".
-  if (result.pr) return <PrChip pr={result.pr} />
-
-  // No PR: offer Create-PR only on a sensible branch (not the default). On the default
-  // branch there's nothing to render.
-  if (isDefault) return null
-
-  return (
-    <div className="border-b border-border-muted px-3 py-2">
-      {!hasUpstream ? (
-        // The push gate: `gh pr create` non-interactively can't prompt to push an
-        // unpushed branch, and we never push on the user's behalf — so guide them first.
-        <p className="text-[11px] text-muted">Push your branch to GitHub to open a pull request.</p>
-      ) : !creating ? (
-        <Button
-          type="button"
-          variant="ghost"
-          size="xs"
-          onClick={() => {
-            // Prefill the title from the branch name — the common, editable starting point.
-            setTitle(branch)
-            setBody('')
-            setCreateError(null)
-            setCreating(true)
-          }}
-          disabled={busy}
-          title={busy ? 'Agent is working…' : 'Create a pull request for this branch'}
-          className="-mx-2 text-muted hover:text-accent-text"
-        >
-          <GitPullRequest className="size-3.5" aria-hidden />
-          Create PR
-        </Button>
-      ) : (
-        <div className="flex flex-col gap-2">
-          <Input
-            autoFocus
-            aria-label="Pull request title"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-            placeholder="PR title"
-            disabled={busy || submitting}
-            className="h-8 text-[13px]"
-          />
-          <Textarea
-            aria-label="Pull request body"
-            value={body}
-            onChange={(e) => setBody(e.target.value)}
-            placeholder="Description (optional)"
-            rows={2}
-            disabled={busy || submitting}
-            className="min-h-16 resize-y text-[13px]"
-          />
-          {createError && (
-            <p className="text-[11px] text-bad" role="alert">
-              {createError}
-            </p>
-          )}
-          {busy && <p className="text-[11px] text-muted">Agent is working…</p>}
-          <div className="flex items-center gap-2">
-            <Button
-              type="button"
-              size="sm"
-              onClick={() => void createPr()}
-              disabled={busy || submitting || title.trim().length === 0}
-            >
-              {submitting ? 'Creating…' : 'Create PR'}
-            </Button>
-            <Button
-              type="button"
-              variant="ghost"
-              size="sm"
-              onClick={() => {
-                setCreating(false)
-                setCreateError(null)
-              }}
-              disabled={submitting}
-              className="text-muted hover:text-accent-text"
-            >
-              Cancel
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
-  )
-}
-
-/** A PR's chip accent by gh state — open reads positive, merged accent, closed negative. */
-function prStateClass(state: string): string {
-  const s = state.toUpperCase()
-  if (s === 'MERGED') return 'text-accent-text'
-  if (s === 'CLOSED') return 'text-bad'
-  return 'text-ok' // OPEN (and any unknown) reads as live.
-}
-
-/**
- * The compact PR chip: `#<number> · <title>`, state-tinted, as an EXTERNAL anchor. The
- * `target="_blank"` + `rel="noreferrer"` makes Electron route the click through main's
- * `setWindowOpenHandler` -> `shell.openExternal`, opening the PR in the system browser
- * (no new IPC). `href` is the gh-provided PR URL.
- */
-function PrChip({ pr }: { pr: GhPr }): JSX.Element {
-  return (
-    <div className="border-b border-border-muted px-3 py-2">
-      <a
-        href={pr.url}
-        target="_blank"
-        rel="noreferrer"
-        title={`#${pr.number} · ${pr.title} (${pr.state})`}
-        className="flex items-center gap-1.5 rounded-md px-2 py-1 text-[13px] transition-colors hover:bg-accent/10"
-      >
-        <GitPullRequest className={cn('size-3.5 shrink-0', prStateClass(pr.state))} aria-hidden />
-        <span className={cn('shrink-0 font-medium tabular-nums', prStateClass(pr.state))}>#{pr.number}</span>
-        <span className="min-w-0 flex-1 truncate text-text">{pr.title}</span>
-      </a>
-    </div>
-  )
-}
-
-/** A glyph's accent: added/untracked read positive, deleted negative, else neutral. */
-function glyphClass(glyph: string): string {
-  if (glyph === 'A' || glyph === 'U') return 'text-ok'
-  if (glyph === 'D') return 'text-bad'
-  return 'text-accent-text'
-}
-
-/**
- * One changed-file row. A leading checkbox toggles the file's commit selection (#86)
- * WITHOUT opening the diff (it's a separate control, not nested in the row button);
- * clicking the rest of the row opens the file's working-tree diff (#85, DIFF mode).
- * The row insets from the panel edges (mx-2) so its rounded hover tint floats clear of
- * the border, matching the menu/list idiom.
- */
-function FileRow({
-  file,
-  checked,
-  onToggle,
-  onSelect,
-}: {
-  file: GitFileView
-  checked: boolean
-  onToggle: () => void
-  onSelect: () => void
-}): JSX.Element {
-  return (
-    <li className="mx-2 flex items-center gap-1 rounded-md transition-colors hover:bg-accent/10">
-      <input
-        type="checkbox"
-        checked={checked}
-        onChange={onToggle}
-        aria-label={`Include ${file.path} in commit`}
-        title="Include in commit"
-        className="ml-2 size-3.5 shrink-0 rounded accent-accent"
-      />
-      <button
-        type="button"
-        onClick={onSelect}
-        title={file.path}
-        className="flex min-w-0 flex-1 items-center gap-2 px-2 py-1.5 text-left text-[13px]"
-      >
-        <span
-          className={cn('w-3 shrink-0 text-center font-semibold tabular-nums', glyphClass(file.glyph))}
-          title={file.glyphLabel}
-          aria-label={file.glyphLabel}
-        >
-          {file.glyph}
-        </span>
-        <span className="min-w-0 flex-1 truncate" dir="rtl">
-          {file.path}
-        </span>
-        {(file.insertions > 0 || file.deletions > 0) && (
-          <span className="shrink-0 tabular-nums text-[11px]">
-            {file.insertions > 0 && <span className="text-ok">+{file.insertions}</span>}
-            {file.insertions > 0 && file.deletions > 0 && ' '}
-            {file.deletions > 0 && <span className="text-bad">−{file.deletions}</span>}
-          </span>
-        )}
-      </button>
-    </li>
   )
 }
